Rename Exchange blocklist file and remove its test file
